client/requirements.txt
pywebview>=4.0
requests>=2.25
